+++ datacube/index/postgres/_api.py
@@ -769,11 +769,14 @@ class PostgresDbAPI(object):
             ]
 
     def get_archived_locations(self, dataset_id):
+        """
+        Return a list of uris and archived_times for a dataset
+        """
         return [
-            record[0]
-            for record in self._connection.execute(
+            (location_uri, archived_time)
+            for location_uri, archived_time in self._connection.execute(
                 select([
-                    _dataset_uri_field(DATASET_LOCATION)
+                    _dataset_uri_field(DATASET_LOCATION), DATASET_LOCATION.c.archived
                 ]).where(
                     and_(DATASET_LOCATION.c.dataset_ref == dataset_id, DATASET_LOCATION.c.archived != None)
                 ).order_by(