+++ docs/about/whats_new.rst
@@ -27,6 +27,11 @@ Usability Improvements
 
  - Significantly reduced disk space usage for integration tests
 
+ - `Dataset` objects now have an `is_active` field to mirror `is_archived`.
+
+ - Added `index.datasets.get_archived_location_times()` to see when each
+   location was archived.
+
 v1.4.1 (25 May 2017)
 --------------------